requirements.txt
py-xmltv
typer[all]
